Basic organizational expense accounts, with a tester function for assert.
Note that Expense account entries need to be tagged with an Invoice, Receipt,
or Statement tag (those tag declarations will be added in next commit).

payee with "NEVER CHARGED"
==========================

The only exception is when the payee has been modified to indicate that the
expense was "NEVER CHARGED".  This was a special-case we encountered where we
were expecting something, they never invoiced us, and it turns out the thing
to be invoiced for never happened.

Typically, we'd just remove the entry from the Ledger file, and allow the VCS
to log the fact that we thought we'd be charged and weren't.  However, in
this case, the expense had been post-audit and therefore was a permanent
fixture on our books.  Changing the payee allowed it to pass checks.

However, going forward, we'd likely never enter anything the ledger UNTIL we
had real proof via an Invoice, Receipt or Statement that showed the Expense
did/should occur.