Split out config-npo.ledger into four different files.
I realized that it makes more sense, even if it does create extra files, for
the accounts, commodities and tags declarations of the project to be
carefully split into different files.  It's definitely more didactic to have
separate files for these declarations, to note that they are, in fact,
separable.

More to that point, users who cut-and-paste from this project into their own
setup will likely be least interested in the chart of accounts, which is
likely to be the longest file by far.  The tags will probably be the most
interesting, so it's important they are in a stand-alone file that can be
easily found.